\name{pin-package}
\alias{pin-package}
\docType{package}
\title{
PIN
}
\description{
PIN corrects the decline in tree ring stable isotope ratios (d13C)
     over the industrial period (> 1850) resulting from increased
     atmospheric CO2 levels using estimates of pre-industrial values
     for the internal (leaf) CO2 concentration (ci).

}
\details{
\tabular{ll}{
Package: \tab pin\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-11-22\cr
License: \tab GPL>=3\cr
LazyLoad: \tab yes\cr
}
}
\author{
Sietse Los <s.o.los@swansea.ac.uk>

Maintainer: same
}
\references{
    see documentation pin
}
\seealso{
Loess
}
\examples{
# see pin
}
